Refactored PIR tests, splitting modeling tests into separate files.
