+++ tests/test_pir_modeling_destruction.py
@@ -1,0 +1,33 @@
+#!/usr/bin/env python3 -B
+import unittest
+import os
+import os.path
+import hashlib
+import json
+import uuid
+import pprint
+import inspect
+from pathlib import Path
+import warnings
+
+from tests import TestProvenancePipelineOutput
+
+class PIRModelingTest_Destruction(TestProvenancePipelineOutput):
+	def test_modeling_for_destruction(self):
+		'''
+		All objects in this set have a 'destroyed_by' property that has type 'Destruction'.
+		
+		This destruction information comes from either the `lot_notes` or the
+		`present_loc_geog` fields.
+		'''
+		output = self.run_pipeline('destruction')
+
+		objects = output['model-object']
+		self.assertEqual(len(objects), 16)
+		for o in objects.values():
+			self.assertIn('destroyed_by', o)
+			self.assertEqual(o['destroyed_by']['type'], 'Destruction')
+
+
+if __name__ == '__main__':
+	unittest.main()
